requirements.txt
# Ozon Parser API - Dependencies

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Async
asyncio-throttle==1.0.2

# Browser automation
playwright==1.40.0
playwright-stealth==1.0.6

# Google Sheets
gspread==6.0.0
google-auth==2.26.1

# Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Logging
loguru==0.7.2
